Synchronization methods implemented.
Both of mutex and spinlock are implemented on semaphore, based on single
core cpu. All interrupts are disabled before jumping to the vector
table(in fact, it's not. interrupts are disabled in schedule()).

A little usage:

1. If data or region is accessed by more than a task, use mutex_lock()
  It guarantees you access the data exclusively. You go sleep until you
  get the key.
2. If the one you are accessing to is the resource that the system also
manipulates in an interrupt, use spinlock_irqsave().
  spinlock never goes to sleep.
3. In case a task need to go sleep, then use_mutex_lock(). And as soon
as you get the key, disable irq, cli().
  Ensure that irq is not disabled before getting the key so that the
  task can go sleep. If you disable irq first before getting the key you
  never get awaken again.