The place of context switching moved from schedule() to interrupt handler().
Actual context switching happens in schedule_prepare() and
schedule_finish() macros. As current context is saved and all interrupts
get disabled by schedule_prepare(), it is safe to use any registers and
free to all interrupts between the macros. But need to be careful as
much as how greedy it is.