Change word data type from `long` to `int`
Let's focus on 8-bit to 32-bit system.

One thing that I don't know what to do about remained, system call.
As system call is an essential component of operating system, I was
thinking making it built-in not configurable in `CONFIGURE`. But what
makes me hesitating is that it would be expensive to implement in 8-bit
AVR. I can't stop doubt system performance aside from technical issues.