Enter `stop` power saving mode when no jobs left to be done
It is the best chance to enter stop mode when idle task takes place with
no timer running. If there is a timer running it means you have still a
job to be done when the timer expires. In that case you enter sleep
mode.

`debug()` macro changed as well reading the document:
https://gcc.gnu.org/onlinedocs/cpp/Variadic-Macros.html#Variadic-Macros