Make embeded file system and timer functionality modular
Disabling `CONFIG_FS`, `ramfs` gets the root file system and mount
`/dev/` on it. But `/dev/` directory isn't visible to user since it is
not actually created in the root file system. There is calling
dependency making it hard to create `/dev' directory between mounting
`rootfs` and `devfs`.

1. Mount `rootfs` first
2. Create `/dev/` directory
3. Mount `devfs` now

What makes it difficult is that to get [2] done system call
functionality must be activated which only can be activated after
`devfs` mounted.

`embedfs` also need to get optimized as it creates `/dev/` in an
incompatible way, not configuable to change its name.