Remove or replace `dsb()` with `barrier()` where not needed or `barrier()` is enough
	ZOMBIE <--- RUNNING <--- STOPPED
	          //|     |\\     (init state; never get back)
	         //         \\
	       |//           \\|
	    SLEEPING        WAITING

A broken link with state of STOPPED sometimes keep hold in run queue. It
causes infinitive loop there referencing itself for prev and next. It
seems obvious that is because of synchronization. But I have no luck yet
to find out which point should be cared of. Need more work and study.

Test after turning off compiler optimization.