fs: Fix the bug of deleting the inodes while still being used
Don't forget when you have deferred work from a system call that the
child of `clone()` only gets its chance to be run after the parent task
returns back from the system call. That means even if you do `resched()`
immediately in the parent task after `clone()`, it doesn't get
rescheduled at the time but as soon as returning from the system call.