stm32: Fix GPIO configuration error when setting it as alternative function and change usart maximum channel to 3 to save buffer memory
When it deals with GPIO as an alternative function for output on stm32f1
didn't work at all because it was in input mode.

To save even little more memory I have changed the number of the default
usart channels opend. It can be easily changed modifying
`USART_CHANNEL_MAX` in `/arch/armv7-m/mach-stm32/include/usart.h`, 3 is
enough typically though.