Add a lock in the task structure, which must be taken before task's property changes
The lock is to get where `set_task_*` and 'get_task_*` macros take
places.

Note that timer would cause a race condition because the interrupts are
not disabled before getting the lock. A timer instance runs in user
mode, technically the mode it is invoked by. So no permission to control
interrupts. Remove the need of controlling interrupts in a timer
instance, which is to manupulate its mode and priority. Detailed in
`/kernel/timer.c`.