embedfs: Fix the precise data bus error
It was caused by the old data written prior in a newly allocated block.
Deleting data when freeing the block would help the system to be secure
for sure. But it takes way much time getting a zeroed block, which
causes additional flash read, write, and erase as well, buffer-cache
takes place at this point though. So here is the deal: it gets a zeroed
block only for the inode data block table.