Reconstruct first-fit allocator and add heap_t data type
I was thinking about implementing TLSF O(1) allocator. And also looking
for a way to reduce places where the interrupts should be disabled
because of synchronization, causing performance bottleneck. This one
with lock-free implementation will remove the need of disabling
interrupts, but will not provide deterministic response time on the
other hand. I haven't done a test on fragmentation neither.