uart: Remove the static buffer placed at compile time in the driver
The number of uarts had to be determined at the compile time to
calculate and save memory for data buffer and wait queue. It was not
only waste of memory that would not be used at run-time on but also such
a restriction. Now it allocates memory dynamically since device ID can
be retrieved from IRQ number.