timer: Fix deadlock caused by priority inversion
When a timer added from waitqueue it used to deal with two locks for
the task and the queue. So timerd would try to get the lock for the task
infinitely while lower priority task is holding the lock. Removed the
possibility of deadlock locking for the task after adding a timer.