Reconstruct softirq and replace some of daemons in driver with it
Applied few of fixes as well.

The way of delivering deferred jobs is changed. A task for the job is
no more is in residence in memory. But raise softirq then softirq
creates a task dynamically to do the job.

As dynamic portion is growing, memory and response time should be
checked carefully. If too many timer requests come, such temporary tasks
will be created possibly causing memory shortage and latency, for
example. Feel like getting away from realtime further and further.