Add syscall_delegate_atomic() for response time and determinism
`CONFIG_SYSCALL_THREAD` provided to turn on/off the feature. If set
every syscall is run by make() creating new task, doing dynamic memory
allocation and losing determinism as it has to wait to be scheduled as
newly created.

On the other hand, syscall uses the caller resource, stack, no delay
waiting to be scheduled. syscall should not be nested in this case
because it uses more stack as we usually work with small memory in
embedded system.