Add a functionality of changing systick interval to sleep longer
Things to be done:

It is okay when woken up by systick interrupt because it is already set
as periodic in the last update. But the problem arises when it is woken
up by any other interrupts. If an interrupt occurs and tasks take place
from then, each task would have time slice as much as the current
interval which can be up to get_raw_sysclk_max(). So
run_sysclk_periodic() should be placed in every interrupt handler. maybe
I think implement the common interrupt handler to put it in only a
place.