Fix entering stop mode while a user task is still running
`TASK_BACKGROUND` state is added to represent the idle task running with
not `TASK_RUNNING` state. The idle task's state could be changed at any
time since the job of destroying zombie tasks is delegated to it. So
need to distinguish whether background state for power saving or running
state for destroying zombies.